// src/model/enum/verification-stage-status.enum.ts
export enum VerificationStageStatus {
    INITIATED = 'INITIATED',
    PENDING_ACCEPTANCE = 'PENDING_ACCEPTANCE',
    IN_REVIEW = 'IN_REVIEW',
    VERIFICATION_ACCEPTED = 'VERIFICATION_ACCEPTED',
    VERIFICATION_REJECTED = 'VERIFICATION_REJECTED',
    PENDING_PAYMENT = 'PENDING_PAYMENT',
    PAYMENT_VERIFIED = 'PAYMENT_VERIFIED',
    STAGE_1 = 'STAGE_1',
    STAGE_2 = 'STAGE_2',
    STAGE_3 = 'STAGE_3',
    VERIFICATION_COMPLETE = 'VERIFICATION_COMPLETE',
}