// src/model/enum/email-type.enum.ts
export enum EmailType {
  EMAILVERIFICATION = 'EMAIL_VERIFICATION',
  ACCOUNTVERIFICATION = 'ACCOUNT_VERIFICATION',
  ACCOUNTREGISTRATION = 'ACCOUNT_REGISTRATION',
  PASSWORDRESET = 'PASSWORD_RESET',
  CONTACTMEREPLY = 'CONTACTMEREPLY',
  ADMINCONTACTMEREPLY = 'ADMINCONTACTMEREPLY',
  COMPANY_VERIFIED = 'COMPANY_VERIFIED',
  COMPANY_REJECTED = 'COMPANY_REJECTED',
  PROPERTY_VERIFIED = 'PROPERTY_VERIFIED',
  PROPERTY_REJECTED = 'PROPERTY_REJECTED',
  COMPANY_SUBMITTED = 'COMPANY_SUBMITTED',
  PROPERTY_SUBMITTED = 'PROPERTY_SUBMITTED',
  VERIFICATION_PIPELINE_UPDATE = 'VERIFICATION_PIPELINE_UPDATE',
  PAYMENT_RECEIPT = 'PAYMENT_RECEIPT',
}